+++ v2/games/packet-manager.h
@@ -1,6 +1,7 @@
 #ifndef PACKET_MANAGER_H
 #define PACKET_MANAGER_H
 #include "games.h"
+#include <string.h>
 
 class PacketManager {
 private:
@@ -11,7 +12,7 @@ public:
   //Constructeur
   PacketManager(Games *currentGame);
   //Permet de gérer un paquet reçu, vérifier sa conformité, l'enregistrer et l'envoyer vers les queues
-  void handlePacket(char* rawPacket, int *rawPacketSize);
+  void handlePacket(char* rawPacket, ssize_t *rawPacketSize);
   //Destructeur
   ~PacketManager();
 };
